Psychotherapist sidebar implementation incomplete
