+++ app/psychotherapist/pages/clientpayment/page.tsx
@@ -1,0 +1,18 @@
+'use client';
+
+import React from "react";
+import { useRouter } from 'next/navigation'; // for handling navigation
+import Sidebar from "@/psychotherapist/components/SideBar";
+
+// PaymentHistory Component
+const ClientPayment: React.FC = () => {
+  const router = useRouter(); // for navigation
+
+  return (
+    <div>
+      <Sidebar></Sidebar>
+    </div>
+  );
+};
+
+export default ClientPayment;
