+++ app/psychotherapist/pages/resources/page.tsx
